Get rid of duplicated code
